fix config swagger
